Add module evaluation.py.
This module contains the classes evaluation and evaluations which are very similar in structure to the classes system and systems implemented in system.py. The largest difference between the system and evaluation classes lays in the class specific attributes and the evaluation class' implementation of its own read, _read, and _from_configs methods; overriding the methods present in the ancestor Configurable.

As of this commit the evaluation classes, have little functionality other then the implementation of the base structure by which configuration information can be used to initialize a class with the relevant attributes.