Create the prepare_data method.
This method mutates the new attribute data which replaces results. The attribute _data stores the unmutated values loaded from the database, allowing for a quick change of class attributes and a new evaluation beginning with a second call of prepare_data (avoiding have to load the data again from the database.