Seperate attribute mutation and metric calculation.
Until now methods such as perform_metrics and summarize performed both the taks of calculating the metrics and summaries based on the input data, but also correctly transferring and formatting this data to their respective attributes (self.evaluation, self.kpi). To clarify the function of the method and provide readable code these methods now return output which is then transferred to the attributes in question in the method run.

Further Notes:

Now it is no longer possible to pass more than one summary per metric, the only valide speration character in the configs is a comma.

The counting of the number of points in each group has temporarily been disabled, due avoid redundant columns when multiple metrics appear; this will be readded in the future.