Make the Evaluations class the central object of the evaluation module.
The Evaluations object is now the central object of the evaluations.py module, which entails two main changes. The first of these changes is to make Evaluations the child of Configurable, opposed to Evaluation. This change is mainly aesthetic, and serves the purpose of highlighting that filling out the fields of the configuration file corresponds to the configuration of the Evaluations object (containing a number of Evaluation instances, corresponding to each section of the config file). The second change is for the load_results method to be moved from the Evaluation class to the Evaluations class. Now, each Evaluation instance belonging to an instance of the class Evaluations fetches its data from Evaluations._data, opposed to loading the data itself from the database indicated in the configuration file.